css finished?
spent the last 2 days playing around with the CSS and I think it should be okay.

I still need to make code for getting between the settings pages. But I believe that the CSS should be good, so it's just a matter of actually accessing them. But we'll see.

also changed some comments just to make them more straightforward